Better handling of optional dependencies
